+++ module/item-sheet.js
@@ -43,7 +43,9 @@ export class GurpsItemSheet extends ItemSheet {
         'data.equipped': true,
         'data.bonuses': `DX+10
 S:Fast*+20
-iq+30`})
+iq+30
+PER+1
+ST-8`})
 }
 )